Create README.md
I made a tik-tak-toe using html5, css3, js (jquery) to test my knowledge in the general front end application and firm my knowledge with jquery. And the result was very cool, a well-done, responsive and fun project came out.

I worked with arrays, lists, objects, functions and fixed jquery. The next objective of the game is to develop a robot that plays against the person.